# YouTooAPI_V2/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "YouTooAPI_V2",
    "version": "v1"
  },
  "paths": {
    "/Conversations/{conversationId}": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponseList"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponseList"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponseList"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Conversations"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IActionResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IActionResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IActionResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Conversations": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFilterBaseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFilterBaseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFilterBaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationOutputModelBaseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Conversations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Conversations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationInputModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32BaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationFilter": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ConversationFilterBaseRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "filter": {
            "$ref": "#/components/schemas/ConversationFilter"
          }
        },
        "additionalProperties": false
      },
      "ConversationInputModel": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "groupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "isAdministrative": {
            "type": "boolean"
          },
          "mutedTo": {
            "type": "string",
            "format": "date-time"
          },
          "conversationStatus": {
            "$ref": "#/components/schemas/ConversationStatus"
          },
          "isPined": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ConversationOutputModel": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "isAdministrative": {
            "type": "boolean"
          },
          "mutedTo": {
            "type": "string",
            "format": "date-time"
          },
          "conversationStatus": {
            "$ref": "#/components/schemas/ConversationStatus"
          },
          "isPined": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ConversationOutputModelBaseResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ConversationOutputModel"
          }
        },
        "additionalProperties": false
      },
      "ConversationOutputModelBaseResponseList": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationOutputModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConversationStatus": {
        "enum": [
          0,
          1,
          5,
          10,
          15,
          20
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
      "Int32BaseResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "data": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}